installer: offer to enable symbolic links
Symbolic links are available since Windows Vista (and since v2.10.1, Git
for Windows only supports Vista and later). While these symbolic links
have *slightly* different semantics than the POSIX ones, they are similar
enough.

With one exception: permission to create symbolic links is only granted to
administrators by default.

However, it is relatively easy to grant that permission, as detailed here:

	https://github.com/git-for-windows/git/wiki/Symbolic-Links

Let's try to detect if permission has been granted at least to the
(non-admin) account used to start the installer, and enable symlinks by
default if that user is permitted to create them.

Note: technically, we do not need to do *anything* in the installer. Why?
Because the `core.symlinks` flag is set upon `git init` (or `git clone`)
when Git detects that it cannot create symbolic links. We choose to show
the option in the installer, though, to increase the visibility both of
the feature as well as of the wiki page describing possible problems with
it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>