installer: report post-install errors more verbosely
There had been two or three reports of the post-install script failing,
but we did not have any better information because the output of the
script was suppressed.

Let's redirect the output to a file and include the contents in any error
report so that we have a chance to help users better in the future.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>