installer (silent): optionally skip installation when files are in use
When the Git Bash is running, for example, Git for Windows cannot be
upgraded. Let's introduce a flag (/SKIPIFINUSE=1) that tells the installer
to please just exit, pretending to have succeeded, if we cannot upgrade
because files are in use. This flag only makes sense in silent
installation mode, of course, as the user would otherwise have a chance to
cancel the installation herself.

This option is also in effect when the installer was called from the VS
installer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>