installer: exit nicely when debugging
For debugging, we introduced a command-line option that allows to build
a "mini" installer that does not actually bundle any files, but only
shows the specified wizard page.

We used to simply call `Abort()` after said page was shown, but that left
`%TEMP%\is-*.tmp` directories behind. Let's instead call the function
`ExitEarlyWithSuccess()` that we introduced elsewhere, for a different
reason, which cleans up that directory (and sets the exit status to 0,
which is a nice additional touch).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>