please.sh tag_git: report the tag name
It can be a bit confusing when the `tag_git` command finishes and the
actual tag name is nowhere to be seen. It can be even unsettling, as this
maintainer would then worry that the automatic naming somehow messed up.

Let's just report the tag name and put worriers at ease.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>