installer: allow to skip downgrades without failing
When installing Git for Windows as part of a larger software suite, we
may be interested in just making sure that Git for Windows is installed,
being happy with newer versions, too. In such a case, we will want to
skip downgrading Git for Windows, indicating success by setting the exit
status to 0.

This behavior needs to be triggered specifically by passing the
/SKIPDOWNGRADE=1 option to the installer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>