installer: hot-fix for post-install "fix"
A recent patch changed the installer to report failures of the
post-install script. However, this patch unfortunately triggered also
when no failure happened (because the post-install.bat script deletes
itself, which leads to cmd.exe erroring out).

Let's try to be extra careful to test whether the batch file has been
removed successfully, and be quiet if it is gone.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>