please.sh prerelease: be more correct about skipping the build
We try to be smart about rebuilding the mingw-w64-git packages if we
find that they already have been built: the build takes a long time.

However, it is not enough to find packages that *seem* to have the
correct version number: we also need to make sure that the tag still
points to the same revision.

If it does not, we *need* to rebuild the packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>