please.sh: fix 'require' invocations
The 'require' function takes a package name, or a group name followed by
the name of a tell-tale package (i.e. if that package is installed, we
assume the entire group was installed).

This patch fixes a couple of unnecessarily repeated arguments, and it
also fixes the issue where the i686 toolchain was *always* installed but
the x86_64 toolchain possibly forgotten.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>