please.sh sync: handle .pacnew files
When Pacman wants to update files that are locally modified, it usually
overwrites the local modifications without remorse. There is a way to
mark files as "backup files" in the package definition, which tells
Pacman to *not* overwrite locally-modified files but write .pacnew files
instead.

Unfortunately, we have to use a very dirty way to override some of
MSYS2's defaults: we edit files that are part of some Pacman packages,
via git-extra's post-install script.

Some of the edited files are marked as backup files, though. So we may
end up with a file `/etc/profile.pacnew` that has all the modifications
delivered by a new version of the `filesystem` package, but the
`/etc/profile` file has only our edits.

To counter that, we need to inspect `/var/log/pacman.log` to detect this
situation, and force-rename the `.pacnew` files before giving
git-extra's post-install script another chance to edit the config files.

This closes https://github.com/git-for-windows/git/issues/951

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>