Revert the latest three commits
They were all misguided attempts at trying to fix the failures of this
developer's build agent to build some prereleases.

The symptom was that the `asciidoctor` executable could not be found, and
the immediate reaction was to try to fix the PATH isues because the
executable was clearly still there.

Except that it was not the correct executable: when building the Git
packages, we build both 32-bit and 64-bit versions. And for the 32-bit
versions, we prefix the PATH with /mingw32/bin/ (and remove any mentions
of /mingw64/bin/).

The real fix was in installing 32-bit asciidoctor by calling

	/mingw32/bin/gem.exe install asciidoctor

This should have been done automatically as part of installing
mingw-w64-i686-asciidoctor-extensions, but was most likely blocked by the
firewall because gem.exe is not allowed by default to access the big wide
internet.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>