please.sh prerelease: when installing Git, overwrite files if necessary
The primary audience of the prerelease subcommand is Continuous
Integration. Such setups are a little bit special, in that their SDKs
may be synchronized between build agents for efficiency, and the
timestamps of the files may not be preserved as Pacman expects them to
be preserved. When that happens, Pacman sees conflicts where there are
none.

To avoid problems in that situation, simply force-install the Git
packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>