please.sh prerelease --force-version: fix %(infix:...)
Contrary to what this developer thought, we actually *cannot* use the
output of `git describe` as tag name: the -g<short-sha1> suffix is
interpreted by `git rev-parse` as "this is the SHA-1 of the actual
revision". As a consequence, the test whether a tag of this name already
exists always returned true.

Let's just replace the dash in the suffix by a period. Period.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>