please.sh prerelease: download previous package if necessary
When pre-releasing, we want to roll back the Git packages after
installing temporary versions. To that end, we need to have the package
files in the first place.

Previously, we used the fall-back to copy the package file from
/usr/src/MINGW-packages/mingw-w64-git/, under the assumption that the
package file must have been built manually if it was not installed from
a Pacman repository.

However, there is also the possibility that the package was installed
elsewhere and the SDK files were copied (without the cached package
files) to a build agent, which now lacks the package file. In this case,
simply download the file from the Pacman repository and be done with it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>