please.sh: avoid PATH problems with `sh -l` invocations
MSYS2 recently added support for a feature where you can limit the PATH,
or inherit it. In Git for Windows we always want to inherit it, e.g. to
keep PATH elements added by other software.

Additionally, the MSYS2_PATH_TYPE=inherit code path supports overriding
the PATH via setting the ORIGINAL_PATH environment variable.

There is a subtle bug in MSYS2's inheritance scheme, though: it
sets the ORIGINAL_PATH environment variable and leaves it set. When
launching a non-MSYS2 program, this environment variable is of course
converted from a Unix-style path list to a Windows one. And when that
non-MSYS2 program happens to call an MSYS2 program again, that
Windows-style path list is not converted back (only "PATH" is, not
"ORIGINAL_PATH").

This hurts us in the all-purpose administrativa script "please.sh" that is
used for all kinds of maintenance tasks and build jobs: It calls
git-cmd.exe in order to leave the MSYS2 realm (so that differing
msys-2.0.dll versions do not interfere if please.sh was called from a
non-SDK), with --command=usr\bin\sh.exe to enter the MSYS2 realm again.

As a work-around, let's just unset the ORIGINAL_PATH variable, to force
MSYS2 to fall back to inheriting the current PATH (which was converted
correctly).

The symptom was a build job that failed to find `attrib.exe` in t0001 and
t5611.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>