please.sh prerelease: add code to upload immediately
The primary reason to introduce the `prerelease` command was to allow
this here maintainer to publish intermediate installers for a wider
testing audience.

It already simplified things greatly, but there is some ways to go:
publishing a prerelease should be as easy as calling

	/path/to/please.sh prerelease --now HEAD

in a Git checkout, and it should build installers, portable Git and
MinGit and upload them all to a specially-crafted tag.

This change implements the uploading part, but does not offer an option
to make use of it yet.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>