please.sh prerelease: allow automatic prerelease tagging
With the %(prerelease-tag) placeholder, --force-version now generates a
sensible version number (incrementing the last digit of the base
version, and appending an auto-incrementing counter).

For example, when generating a version for a revision that is based on
v2.11.0.windows.1, it would result in v2.11.1.windows-prerelease.1 the
first time, v2.11.1.windows-prerelease.2 the second time, etc.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>