please.sh prerelease: automate the entire prerelease & upload process
When publishing a prerelease, it is easy to forget something or get
something else wrong. For example, the version number produced by `git
version` should reflect the prerelease's official version number.

And it should also all be as automatic as possible, so that this here
maintainer simply calls:

	/usr/src/build-extra/please.sh prerelease --now HEAD

in a Git checkout and it will do everything up to, and including,
uploading the artifacts.

This introduces that option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>