please.sh bisect_broken_test: optionally publish a comment on GitHub
If one of Git for Windows' build agents identifies a problem and tells
nobody about it, it is as if it never happened.

So let's do this instead: when we identify a certain commit that broke
tests on Windows, let's just run that test in verbose mode with that
commit and then attach the output as a commit comment on GitHub.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>