please.sh prerelease: do not reuse tag_name multiple times
To avoid proliferation of variable names, this developer tried to reuse
the same variable for multiple purposes. Unfortunately, that failed
spectacularly, as the tag_name variable now tried to serve two different
purposes at the same time: hold the output of `git describe` *and* hold
the replacement for the infix specified in the placeholder
%(infix:<pattern>=<replacement>).

Let's just use the variable name `desc` for the output of `git
describe`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>