please.sh sync: re-install asciidoctor when needed
The asciidoctor package is actually a Ruby gem, not a pacman package.
Therefore, when pacman updates Ruby, it has no idea that some old gems
need to be re-installed for the new Ruby version.

With this patch, the script tests specificially for that situation and
simply re-installs the asciidoctor gem as needed.

Note: the asciidoctor gem is only needed to build Git's documentation,
and as we build both 32-bit and 64-bit Git packages in the 64-bit SDK,
we can skip this step in the 32-bit SDK.

Noticed by Eric Vera.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>