Rename git-lfs/ to reflect that it is a MINGW package
The difference between MSYS2 packages and MINGW ones is that the former
have a (potentially transitive) dependency on the MSYS2 runtime.

As git-lfs does not depend on the MSYS2 POSIX emulation layer, it is
squarely in the MINGW realm, therefore its package name has the
`mingw-w64-` prefix, and hence the directory name should reflect the
same.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>