mingw-w64-git-lfs: include actual source code in the src package
As with the Git Credential Manager, Git for Windows does not actually
build Git-LFS itself but relies on trusted upstream releases. That means
that the `PKGBUILD` file's `source` section actually refers to the
binary packages.

Let's use the same trick as with the Git Credential Manager to include
the actual source code in the src package: simply include the source
archive in the `source` section, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>