Adding rsync to installer
This uses the same method as pull #57 and solves the (closed) #347.

Rationale: I am looking to make backups using rsync, and would like to include it in Git Bash, rather than install other software such as Cygwin. Per #347 there is demand for it. As listed below, the file size increase is negligible (<0.4%).

64-bit compiled version sizes (via `wc -c`):

43508495 Standard (43.51MB)
43678759 With rsync (43.68MB)

Installing rsync via pacman was a 0.25MiB download and 0.46MiB installed, so the 0.17MB difference in the installer checks out.

Signed-off-by: Scott Stevens <go.scott.mac@gmail.com>