please.sh bisect_broken_test: handle trivial case
There is one case that is very trivial: if the "good" commit is the only
parent of the "bad" commit. In this case, there is only one candidate.

Also in this case, "git bisect start" does not even produce a log. Which
is bad in and of itself ;-)

Let's handle this case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>