git-extra: fix leaking of stupidly-chosen "PERL_PATH" variable
In their infinite wisdom, the MSYS2 developers decided to use the
variable name PERL_PATH to define the elements to add to the PATH for
custom Perl scripts, and not only that, but then also to export it to
each and every shell process.

The variable name PERL_PATH has long been established for use cases
where the user wants to refer to the Perl *interpreter*, i.e.
/usr/bin/perl.exe.

This is not just a stupid mistake but also breaks at least Git's
build/test procedure.

Let's fix this as we fix so many other things: by overriding the faulty
code via git-extra's `install` script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>