please.sh sync: fix .pacnew handling
We have this long-standing, annoying problem where MSYS2 keeps adding
stuff that interferes with Git for Windows' operation (such as setting
and exporting a `PERL_PATH` variable containing a list of PATH elements
to add for use with custom Perl scripts, when it has been a
long-standing tradition to assign the location of the Perl interpreter
du jour to that variable instead).

Our strategy is to keep adding fixup commands to git-extra's .install
script that keep editing files installed by other packages.

Some of those files are marked "precious", such as /etc/profile. That
means that pacman will not overwrite modified versions, but instead
write out .pacnew files with the new versions.

What we need to do in that case is to force-rename those .pacnew files
and simply reinstall git-extra.

The code that was trying to detect what .pacnew files pacman decided to
write *this* time was correct for a long time.

But now pacman requires *two* upgrade runs, from time to time, because
the strategy chosen by the MSYS2 project to overcome the problem where
pacman cannot overwrite its own depdendencies (e.g. msys-2.0.dll) and
continue is to simply refuse to stop, asking the user to interactively
close the window and restart the process. The `sync` command of the
please.sh script deals with this issue by detecting whether the log
shows that a full upgrade (as opposed to the "system" upgrade) was
attempted, and simply reissuing the upgrade command.

With this new strategy, the pacman log will contain a "starting system
upgrade" followed by another "starting system upgrade" that is then
followed by a "starting full upgrade.

Hence the search for the .pacnew files cannot simply throw away
previously found results when encountering the "starting system upgrade"
line: it may be the one of the second pacman run, while the first one
may have upgraded the /etc/profile file in the filesystems package.

Instead, the search needs to make sure that it saw a "starting full
upgrade" line before the "starting system upgrade" line before throwing
away previously-found .pacnew results.

This requires quite a bit more sed magic than this developer would like
to employ, but so be it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>