installer: exclude `curl` from being bundled
We actually shipped *two* curl.exe files: one in /usr/bin/ (that depends
on the MSYS2 runtime) and one in /mingw64/bin/ (that does not depend on
the MSYS2 runtime).

According to `pactree -r curl`, the only users of the `curl` package are
- pacman
- pkgfile
- git-flow (via git)

The `git-flow` package depends on `git` (the MSYS2 version, which we
exclude from the installer) only because it is a bunch of shell scripts
and relies on the MSYS2 bash anyway. Therefore, `git-flow` does not, in
fact, require the MSYS2 version of the curl package.

Please note that this analysis concerns the *curl* package, *not* the
*libcurl* package (which is a dependency of gnupg, hence required by Git
for Windows).

Let's just exclude the `curl` package explicitly from the packages whose
contents are bundled in Git for Windows' installer.

This fixes https://github.com/git-for-windows/git/issues/1069.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>