please.sh build: handle MSYSTEM=MSYS properly
The purpose of git-cmd.exe is to set things up for regular Git for
Windows, including setting MSYSTEM=MINGW64 (or MINGW32 for 32-bit
setups).

Naturally, this conflicts with this script's use of git-cmd.exe (which
makes it possible to jump the boundary from a Git for Windows
installation to the Git for Windows SDK, which may very well differ in
their respective msys-2.0.dll).

Previously, we already re-set MSYSTEM=MSYS in the command-line passed to
git-cmd.exe, and re-set the PATH, but there are now more things in
MSYS2's startup scripts depending on the value of MSYSTEM.

Let's just try to force a reinitialisation by sourcing /etc/profile
again. To prevent the common problems with ORIGINAL_PATH, we simply
unset it before.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>