please.sh: add optional "commit" argument to test_remote_branch
If test_remote_branch is triggered from TravisCI then the branch name
might not be sufficient to identify the commit that TravisCI is
checking. Teach test_remote_branch an optional argument to specify
the exact commit hash that should be build and test. This commit hash
needs to be on the remote-tracking-branch that is specified as
mandatory argument.

Signed-off-by: Lars Schneider <larsxschneider@gmail.com>