Add a new MINGW to allow generating .pdb files for Git's executables
The cv2pdb project provides executables that extract the DWARF symbols
from Code View executables (or executables compiled with GCC) and puts
them into .pdb files for use with Visual Studio (or crash dumps).

This is extremely useful when trying to make heads and tails of crash
reports.

As cv2pdb was developed using Visual Studio, there is currently no easy
way to build the executable with GCC, so we simply require Visual Studio
(which is not a big deal because even the free Visual Studio Community
can be used).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>