MinGit: move all core-git .exe files into bin/
Core Git's .exe files may link to .dll files that are already in bin/,
and the way Windows works, it is best to have those .exe and .dll files
in the same directory lest some enterprisey's installer that installs
an incompatible libeay32.dll into %SYSTEMROOT%\system32 interfere with
the functionality.

Note: we *only* move core Git's .exe files, *not* other .exe files in
libexec/git-core/, as they come with their own .dll files that are
already in the same directory.

This fixes https://github.com/git-for-windows/git/issues/1086

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>