mingit: include awk.exe's ncurses dependency
For some really funny reason, awk.exe depends on ncurses: awk features
an interactive console backed by libreadline, and that library uses
ncurses as termcap library in MSYS2.

However, we specifically excluded the `ncurses` package from MinGit,
which means that awk.exe fails to even start.

Now, that would not be a problem if Git itself was not using awk. But
it does: the `git mergetool` command makes use of this helper.

So let's include the bare minimum to make awk.exe work, until the time
when the mergetool command finally becomes a builtin.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>