get-sources: account for the fact that we no longer ship /usr/bin/curl.exe
We exclude the `curl` package, but not the `libcurl` package (needed by
`gnupg`), but both are built from /usr/src/MSYS2-packages/curl/PKGBUILD.

Previously, we relied on the fact that the base package (i.e. the
package with the same name as the subdirectory in MSYS2-packages) is
included with secondary packages, but this is now no longer true.

Let's address that problem generically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>