get-sources: fall back to try downloading from BinTray
The Git for Windows project sometimes provides out-of-band releases of
some of the packages usually shipped by the MSYS2 project.

An example is cURL, which Git for Windows follows much more closely than
MSYS2, for obvious reasons: Git is heavily affected by any bugs in cURL.

So let's just fall back to downloading sources from BinTray even for
packages that are shipped by MSYS2, in case Git for Windows was faster.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>