get-sources: handle duplicates gracefully
The git-doc-html package is a secondary package of the git package, and
as it is listed separately in package-versions.txt, this means that we
try to find the .zip file in the previous zipdir *twice* (but the first
time, we moved it).

Let's just also look in the current zipdir whether we have the .zip and
if we do, take a shortcut to the next entry of package-versions.txt.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>