portable: ship with all .dll files copied from bin -> libexec/git-core
When installing .exe files into a different directory than their .dll
dependencies, there is always a chance that a mismatching version of
the same .dll in C:\WINDOWS\system32 is used.

For that reason, we already have a post install script in place that
tries to hardlink all .dll files from the bin directory into the
libexec/git-core directory, falling back to copying them.

However, this only works if the portable Git was installed using the
self-extracting installer that executes that post install script, or
if Git Bash was started at least once, or if the post install script
was executed manually.

We should do better than this: start with copied versions, and try to
hard-link in the post install script if possible.

This addresses https://github.com/git-for-windows/git/issues/1086

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>