please.sh: add support for handling the p7zip package
We just switched the portable Git to include the `.dll` files also in
libexec/git-core in addition to the bin directory. However, the p7zip
version that is shipped with standard MSYS2 does not do a good job on
those identical files, adding a whopping 4MB to the archive.

The current p7zip version does not have that problem, so let's just
upgrade to that version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>