please.sh prerelease: fix %(infix), once again
When `git describe`'s output ends in ".g<short-hash>", it is actually
reporting an exact tag (which we generated earlier from the output of
another `git describe` call that ended in "-g<short-hash>", i.e. using
a "-" instead of a ".").

So we still want to reject that.

Noticed by Alejandro Pauly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>