please.sh test_remote_branch: do not ignore test failures
For just over a week now, we have code that filters the output of the
tests by default. Which is nice.

However, that code changed the behavior: failed tests are now ignored, as
the exit code of a command is ignored if its output is piped into another
command. Which is not nice.

Work around this issue by recreating the exit code based on the presence
of any "failed <n>" lines in the test-results/ (which are written out by
every full test run).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>