Update the 7-Zip SFX component
For years, Git for Windows has used the "modified SFX" component that
was offered at http://7zsfx.info, which is now sadly defunct.

The reason why we needed to use this component over the "official" 7-Zip
SFX component is that the latter insists on unpacking the entire archive
into a temporary directory, launch the command line specified by the
`RunProgram` config directive (falling back to `setup.exe` in the
top-level directory), *and then delete the temporary directory*.

In contrast, portable Git needs to extract itself into a target
location specified by the user, and then only run a simply
post-install.bat script afterwards.

The "modified SFX" supported this requirement, and others.

The source code for this SFX can be found here:

	https://github.com/git-for-windows/7-Zip

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>