please.sh rebase: only collect garbage after the rebase
Garbage collection tends to take a long time, especially when it is
triggered over and over and over again due to too many loose objects (in
which case Git complains that `git prune` should be run but does not do
that automatically).

So let's just call `gc` and `prune` explicitly after the rebase, and
avoid letting `gc --auto` interfere with the rebase run itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>