installer: limit packsize to 2GB only within 32-bit setups
It is quite possible to install a 32-bit Git for Windows into a 64-bit
Windows, and a portable Git will still read from %PROGRAMDATA%\Git\config.
Therefore it is a mistake to set the packsize globally. So let's not.
Let's do it in the "system-wide" Git config that is private to the
installed 32-bit Git for Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>