installer: configure http.sslCAInfo in "system-wide" config
Git for Windows has that very special thing that the "system-wide" config
actually only extends to the current Git for Windows in use.

The reason: Git's design assumes a Unix environment where there is a
single top-level `/etc/` directory.

This is not true for Windows, where there simply is no equivalent for the
`/etc/` directory.

Instead, we have a `<GIT>\mingw64\etc\gitconfig` that is used by the
`git.exe` installed into the same location.

We *also* have a `C:\ProgramData\Git\config` whose purpose is to hold
configuration that extends to more than just Git for Windows: JGit,
libgit2-based software, portable Git all read from there (or at least are
supposed to).

We have to be careful not to write anything into the ProgramData config
that is private to Git for Windows. And the SSL certificates that are used
by Git for Windows' own copy of cURL is pretty private. So let's configure
it in Git for Windows' "system-wide" config instead of in ProgramData.

This fixes https://github.com/git-for-windows/git/issues/531

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>