wingit-snapshot-helper: implement convenience function to upload a snapshot
The entire idea of starting this script was to make it possible for some
build job to not only build the installers and portable Gits, but to
upload them to a publicly visible place, too.

This patch implements that functionality, in particular the part where
the index.html file is updated, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>