please.sh: skip UNC tests if //localhost/c$ inaccessible
In Docker Windows containers, it seems that the administrative shares (or
any shares, for that matter), are not yet supported.

As we are about to switch to VSTS build agents hosted in Docker containers
(for better management, isolation and scalability), we need to handle this
situation gracefully.

Sadly, the test in t5580 that tries to verify specifically whether the UNC
path is accessible only tests whether it is a directory, and that test
somehow succeeds (probably an unfortunate bug in the MSYS2 runtime when
trying to stat administrative shares).

Therefore, let's just test whether the administrative share can be *read*,
and skip the test otherwise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>