please.sh rebase: allow skipping mode-only changes
There is a patch in current Git for Windows (003f615b6d7 (name-hash: add
perf test for lazy_init_name_hash, 2017-03-23)) that differs only in the
file mode from a patch that made it into upstream's `pu` branch.

Let's handle such a patch gracefully when trying to skip it from the
rebase.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>