please.sh: add a helper to work with Coverity
Coverity is a pretty complete static analysis tool. Working with it is a
bit tricky, as a so-called "Self-Scan" tool needs to be downloaded, and a
full build needs to be run through that tool, the output needs to be
packages into a .tar file and uploaded, queing work for the server side of
Coverity.

To make it simpler to use, let's pour all of our knowledge about the
Coverity process into Git for Windows' administrative script, both to
document it and to make it easier to use.

So now you only need to set the config variables coverity.username and
coverity.token and then you can call

	/usr/src/build-extra/please.sh submit_build_to_coverity \
		--worktree=. HEAD

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>