please.sh: avoid false positives with strbuf & Coverity
Coverity has this annoying behavior where it cannot figure out that our
strbuf_slopbuf is not even used when we grow the strbuf. Hence it
reports gazillions of bogus buffer overruns, which makes the tool
substantially less useful (as it drowns out the *real* bugs).

Work around this by using a real big slopbuf, just for Coverity's sake.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>