installer: no longer try to "bind" to current .dll functions
There is a feature in Windows where you can update the part of the .exe
file which references the address of the functions loaded from .dll
files. This feature nominally speeds up the startup times.

However, updating the .exe files invalidates any cryptographic
signature. What is worse: the bound addresses will almost certainly be
incorrect when using Address Space Layout Randomization (ASLR), which we
use to fend off stack-based attacks preemptively.

So let's just skip that step and leave the .exe files as-are.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>