Add a new package: git-for-windows-keyring
Back in October 2015, this maintainer already tried to sign the Pacman
packages uploaded to Git for Windows' Pacman repository. The ticket to
track the wish, and the progress, can be found here:

	https://github.com/git-for-windows/git/issues/456

The result was a disaster: nobody could update anymore.

Contrary to what this maintainer had been led to believe, it was not
enough to just sign the packages. Pacman's configuration allows for
importing keyrings, and such a keyring needed to be installed *before*
uploading a signed package database (or signed packages).

This commit adds the definition for a Pacman package containing such a
keyring.

Obviously, this adds a Catch-22: the git-for-windows-keyring package
*needs* to be installed before any signed packages (or package
databases) from Git for Windows' Pacman repository can be installed.

Therefore, we will need to run with unsigned packages for a while
longer, maybe something around two months. In the meantime, we need to
update the git-sdk-installer, and the git-sdk-32 and git-sdk-64
repositories in the https://github.com/git-for-windows org.

Hopefully all users of Git for Windows' SDK will have updated by then,
so that we will finally be able to switch on package signing at that
stage.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>