pacman-helper: add & upload signatures if GPGKEY is set
Pacman's repo-add helper already uses that environment variable to let
the user specify which GPG key to use.

Let's take the presence of that environment variable as a strong hint
that the user wants to sign the packages and package databases when
using the pacman-helper.

And when there are .sig files present that correspond to packages to be
uploaded, let's upload them, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>