please.sh test-remote-branch: ignore bogus Travis triggers
Travis seems to be unable to stop some bogus PRs to slip through to the
CI-only jobs.

As a consequence, we see quite a lot of failed builds just because the
commit in question is not on the branch Travis claims it should be on.

Since the primary user of the "commit-to-test" feature is said Travis
job, let's just ignore those bogusly triggered builds.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>