please.sh publish: verify that we have a usable node.exe
There is sadly a big problem with the MSYS2 project, affecting us rather
badly from time to time: every once in a while, a library is upgraded
from one major version to the next, and this version number is
*hardcoded* in the .dll's file name. This means that *all* .exe files
which were happily consuming the previous version now fail to find the
library and error out with exit code 127 instead.

This happened already for the second time with node.exe, when
mingw-w64-icu was upgraded to another major version, and
mingw-w64-nodejs was not rebuilt as required (this issue has been
reported at https://github.com/Alexpux/MINGW-packages/issues/2585, the
underlying bug being that the MSYS2 project has no infrastructure in
place to automatically rebuild packages upon major version bumps of
dependencies, see https://github.com/Alexpux/MINGW-packages/issues/2145
for details, with little hope of an eventual fix).

In the meantime, let's just error out if we know that the update of the
website would fail. The workaround is to rebuild the
mingw-w64-x86_64-nodejs package locally in an up-to-date Git for Windows
SDK, which will make node.exe pick up the latest libicu.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>