please.sh: add a command to automate upgrading packages
It is a recurring maintenance task to build packages of dependencies
(such as Git Credential Manager) when a new version comes out.

Let's automate this, not only to lessen the maintenance burden of Git
for Windows, but also to document the process.

By necessity, this is a process that is pretty specific to the package.
For example, Git Credential Manager's tag sometimes lacks the `v`
prefix, sometimes the zip file name reads "gcmw", sometimes "gcm", and
sometimes the binaries are stored inside a subdirectory in the .zip.

Let's start documenting the package upgrading process with the Git
Credential Manager package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>