vimrc(git-extra): exclude MERGE_EDITMSG from jumping "back"
Git for Windows activates vim's feature that lets it remember the line
where the cursor was when the file was last opened.

This feature makes no sense, of course, for ever-changing files such as
COMMIT_EDITMSG.

Likewise, MERGE_MSG should also be expected to change all the time. This
was suggested by Toby Allsopp in the discussion of
https://github.com/git-for-windows/git/issues/1207.

While at it, mark MERGE_MSG as UTF-8 just like COMMIT_EDITMSG, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>