git-extra: override vim's "jump to last known position" behavior
Since version 8.0, vim has defaults that kick in when there is no
`.vimrc` file in the user's home directory. One of these defaults
makes vim move the cursor when opening files with recorded last cursor
positions.

This behavior is almost what we want in Git for Windows: there are a
couple of files which we do *not* to be treated that way, as the
contents of those files are volatile. For example, commit messages are
always accessed via the same path, but they are rather different almost
every time the user edits them.

We already have code to implement the same feature, excluding files
where we know that it does not make sense to move the cursor "back" to
the recorded position, in `/etc/vimrc`.

Obviously, we do not want vim's defaults to override our carefully
implemented feature. The suggested way of overriding the `vimStartup`
autogroup does *not* work, as `/etc/vimrc` is parsed first, and
`defaults.vim` is parsed afterwards (provided that there is no
`~/.vimrc`).

Another suggested method to override the defaults is by defining
`skip_defaults_vim` in `/etc/vimrc`, but that would override *all* of
the defaults, which is too heavy-handed.

Let's just comment-out that part of vim's `defaults.vim` instead.
so that we can override just that feature.

This fixes https://github.com/git-for-windows/git/issues/1207

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>